requirements.txt
numpy>=1.12
matplotlib
pytorch
pandas 
seaborn 
scikit-learn
PyTorch
Pyro (only for "Pyro" repository)